+++ src/main/java/com/web/SearchWeb/comment/dto/CommentDto.java
@@ -8,8 +8,8 @@ import lombok.ToString;
 @Setter
 @ToString
 public class CommentDto {
-    private int board_boardId;
-    private int member_memberId;
+    private Long board_boardId;
+    private Long member_memberId;
     private String member_nickname;
     private String content;
 }
